Migrate to ros2

+++ src/map_manager.cpp
@@ -646,11 +646,11 @@ void MapManager::addMapPoint(const cv::Scalar &color)
     // Visualization related part for pointcloud obs
     pcl::PointXYZRGB colored_pt;
     if( plm->isobs_ ) {
-        colored_pt = pcl::PointXYZRGB(255, 0, 0);
+        colored_pt = pcl::PointXYZRGB((uint8_t) 255, (uint8_t) 0, (uint8_t) 0);
     } else {
-        colored_pt = pcl::PointXYZRGB(plm->color_[0] 
-                                    , plm->color_[0]
-                                    , plm->color_[0]
+        colored_pt = pcl::PointXYZRGB((float) plm->color_[0] 
+                                    , (float) plm->color_[0]
+                                    , (float) plm->color_[0]
                                     );
     }
     colored_pt.x = 0.;
@@ -674,11 +674,11 @@ void MapManager::addMapPoint(const cv::Mat &desc, const cv::Scalar &color)
     // Visualization related part for pointcloud obs
     pcl::PointXYZRGB colored_pt;
     if( plm->isobs_ ) {
-        colored_pt = pcl::PointXYZRGB(255, 0, 0);
+        colored_pt = pcl::PointXYZRGB((uint8_t) 255, (uint8_t) 0, (uint8_t) 0);
     } else {
-        colored_pt = pcl::PointXYZRGB(plm->color_[0] 
-                                    , plm->color_[0]
-                                    , plm->color_[0]
+        colored_pt = pcl::PointXYZRGB((float) plm->color_[0] 
+                                    , (float) plm->color_[0]
+                                    , (float) plm->color_[0]
                                     );
     }
     colored_pt.x = 0.;
@@ -752,11 +752,11 @@ void MapManager::updateMapPoint(const int lmid, const Eigen::Vector3d &wpt, cons
     // Visualization related part for pointcloud obs
     pcl::PointXYZRGB colored_pt;
     if(plmit->second->isobs_ ) {
-        colored_pt = pcl::PointXYZRGB(255, 0, 0);
+        colored_pt = pcl::PointXYZRGB((uint8_t) 255, (uint8_t) 0, (uint8_t) 0);
     } else {
-        colored_pt = pcl::PointXYZRGB(plmit->second->color_[0] 
-                                    , plmit->second->color_[0]
-                                    , plmit->second->color_[0]
+        colored_pt = pcl::PointXYZRGB((float) plmit->second->color_[0] 
+                                    , (float) plmit->second->color_[0]
+                                    , (float) plmit->second->color_[0]
                                     );
     }
     colored_pt.x = wpt.x();
@@ -874,7 +874,7 @@ void MapManager::mergeMapPoints(const int prevlmid, const int newlmid)
     
     // Visualization related part for pointcloud obs
     pcl::PointXYZRGB colored_pt;
-    colored_pt = pcl::PointXYZRGB(0, 0, 0);
+    colored_pt = pcl::PointXYZRGB((uint8_t) 0, (uint8_t) 0, (uint8_t) 0);
     colored_pt.x = 0.;
     colored_pt.y = 0.;
     colored_pt.z = 0.;
@@ -959,7 +959,7 @@ void MapManager::removeMapPoint(const int lmid)
 
     // Visualization related part for pointcloud obs
     pcl::PointXYZRGB colored_pt;
-    colored_pt = pcl::PointXYZRGB(0, 0, 0);
+    colored_pt = pcl::PointXYZRGB((uint8_t) 0, (uint8_t) 0, (uint8_t) 0);
     colored_pt.x = 0.;
     colored_pt.y = 0.;
     colored_pt.z = 0.;
@@ -1039,9 +1039,9 @@ void MapManager::removeObsFromCurFrameById(const int lmid)
     plmit->second->isobs_ = false;
 
     // Update MP color
-    colored_pt = pcl::PointXYZRGB(plmit->second->color_[0] 
-                                , plmit->second->color_[0]
-                                , plmit->second->color_[0]
+    colored_pt = pcl::PointXYZRGB((float) plmit->second->color_[0] 
+                                , (float) plmit->second->color_[0]
+                                , (float) plmit->second->color_[0]
                                 );
                                 
     colored_pt.x = pcloud_->points.at(lmid).x;
@@ -1071,7 +1071,7 @@ bool MapManager::setMapPointObs(const int lmid)
     plmit->second->isobs_ = true;
 
     // Update MP color
-    colored_pt = pcl::PointXYZRGB(200, 0, 0);
+    colored_pt = pcl::PointXYZRGB((uint8_t) 200, (uint8_t) 0, (uint8_t) 0);
     colored_pt.x = pcloud_->points.at(lmid).x;
     colored_pt.y = pcloud_->points.at(lmid).y;
     colored_pt.z = pcloud_->points.at(lmid).z;
